requirements.txt
google-cloud-bigquery==2.31.0
pandas
pyarrow